Handle errors correctly once
Instead of multiple error handling all over the place, then just do it
in one place.
