+++ cmd/plugins/juju-wait-for/model.go
@@ -110,15 +110,12 @@ func (c *modelCommand) waitFor(name string, deltas []params.Delta, q query.Query
 		case *params.ModelUpdate:
 			if entityInfo.Name == name {
 				scope := MakeModelScope(entityInfo)
-				if res, err := q.BuiltinsRun(scope); query.IsInvalidIdentifierErr(err) {
-					return false, invalidIdentifierError(scope, err)
-				} else if query.IsRuntimeError(err) {
+				if done, err := runQuery(q, scope); err != nil {
 					return false, errors.Trace(err)
-				} else if res && err == nil {
+				} else if done {
 					return true, nil
-				} else if err != nil {
-					logger.Errorf("%v", err)
 				}
+
 				c.found = entityInfo.Life != life.Dead
 			}
 			break
